feat: Rewrite cobalt data layer with @fuman/fetch and implement proxy support
