+++ .env.example
@@ -5,9 +5,10 @@ BOT_TOKEN=
 API_ENDPOINTS='[
     "https://api.cobalt.tools",
     { "url": "https://api.cobalt.tools" },
-    { "name": "custom name", "url": "https://api.cobalt.tools", "auth": "Api-Key 123" }
+    { "name": "custom name", "url": "https://api.cobalt.tools", "auth": "Api-Key 123", "proxy": "http://proxy.example.com" }
 ]'
 
 # Optional variables, defaults shown
 SELECT_TYPE_PHOTO_URL=https://i.otomir23.me/buckets/cobold/download.png
 ERROR_CHAT_ID=
+CUSTOM_INSTANCE_PROXY_URL=
