+++ src/core/data/cobalt/external.ts
@@ -1,0 +1,5 @@
+import { baseFetch } from "@/core/data/cobalt/common"
+
+export async function retrieveExternalMedia(url: string, proxy?: string) {
+    return baseFetch(url, { proxy }).then(r => r.arrayBuffer())
+}
